--- TeamCode/src/main/java/robohawks/controllers/AutoRight.java
@@ -1,55 +1,0 @@
-package robohawks.controllers;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.CRServo;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.Servo;
-import com.qualcomm.robotcore.util.ElapsedTime;
-
-import org.firstinspires.ftc.robotcore.external.ClassFactory;
-import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
-import org.firstinspires.ftc.robotcore.external.tfod.Recognition;
-import org.firstinspires.ftc.robotcore.external.tfod.TFObjectDetector;
-
-import java.util.List;
-
-import robohawks.modules.base.HolonomicDriveModule;
-
-@Autonomous(name = "AutoRight")
-public class AutoRight extends LinearOpMode {
-
-    HolonomicDriveModule drive;
-    ElapsedTime time;
-
-    @Override
-    public void runOpMode() {
-
-        //**********INIT START**********//
-
-        drive = new HolonomicDriveModule(hardwareMap);
-        time = new ElapsedTime();
-        waitForStart();
-
-        //**********INIT STOP**********//
-        //**********DETECT START**********//
-
-        time.reset();
-
-        String pos = "right";
-
-        while (time.seconds() < 0.5 && opModeIsActive()) {
-
-            //RIGHT
-            drive.setPowerOne(-1);
-            drive.setPowerTwo(1);
-            drive.setPowerThree(-1);
-            drive.setPowerFour(1);
-            sleep(1000);
-            drive.setPowerOne(0);
-            drive.setPowerTwo(0);
-            drive.setPowerThree(0);
-            drive.setPowerFour(0);
-        }
-    }
-}
